requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
python-docx==1.1.0
easyocr==1.7.1
torch
torchvision
pytesseract==0.3.13
Pillow==10.4.0
spacy==3.7.5
sumy==0.11.0
nltk==3.9.0
aiofiles==24.1.0
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai
python-dotenv
